remove system calls and migrate thrainio_test
